Rename 1.20.6 module
