+++ settings.gradle.kts
@@ -3,12 +3,12 @@
  */
 
 rootProject.name = "movecraft-parent"
-include(":movecraft-v1_20")
+include(":movecraft-v1_20_6")
 include(":movecraft-v1_21")
 include(":movecraft-api")
 include(":movecraft-datapack")
 include(":movecraft")
-project(":movecraft-v1_20").projectDir = file("v1_20")
+project(":movecraft-v1_20_6").projectDir = file("v1_20_6")
 project(":movecraft-v1_21").projectDir = file("v1_21")
 project(":movecraft-api").projectDir = file("api")
 project(":movecraft-datapack").projectDir = file("datapack")
